+++ internal/acmesolver/constants.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package solver
+package acmesolver
 
 const (
 	// HTTPChallengePath is the path prefix used for http-01 challenge requests
